fix race condition bug
in which shapes are removed from the sandbox before they are added -- this is due to the user clicking too fast, causing touch_down and touch_up events to occur at roughly the same time. now, nothing ends up being drawn to the screen, which is 'safe', and indirectly prompts the user to try again.